Safari-609.1.11.1/LayoutTests/inspector/css/pseudo-creation-expected.txt

Testing when CSS.events.pseudoElementCreated and CSS.events.pseudoElementDestroyed are fired.

Requesting document...

Calling "createElementWithClass("test-pseudo-without-content")"...
Checking for nodes with class ".test-pseudo-without-content"...
PASS: There should be 1 node with the class ".test-pseudo-without-content".

Calling "removeElementWithClass("test-pseudo-without-content")"...
Checking for nodes with class ".test-pseudo-without-content"...
PASS: There should be 0 node with the class ".test-pseudo-without-content".

Calling "createElementWithClass("test-pseudo-with-content")"...
Checking for nodes with class ".test-pseudo-with-content"...
PASS: Created ::before pseudo element
PASS: There should be 1 node with the class ".test-pseudo-with-content".

Calling "removeElementWithClass("test-pseudo-with-content")"...
Checking for nodes with class ".test-pseudo-with-content"...
PASS: Removed ::before pseudo element
PASS: There should be 0 node with the class ".test-pseudo-with-content".
